+++ test/index.js
@@ -1,0 +1,7 @@
+'use strict'
+require('./skip')
+require('./update')
+require('./create')
+require('./logs')
+require('./aggregate')
+require('./merge')